requirements.txt
openai==1.36.1
openpyxl==3.1.2
python-dotenv==1.0.1
selenium==4.18.1
argparse==1.4.0
rapidfuzz